retry

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <time.h>
 #include <stdlib.h>
+
 /**
  * main- Entry point
  *
@@ -8,21 +9,21 @@
  */
 int main(void)
 {
-	int n;
+	long int n;
 
 	srand(time(0));
 	n = (rand() - RAND_MAX) / 2;
 	if (n > 0)
 	{
-		printf("%d is positive\n", n);
+		printf("%ld is positive\n", n);
 	}
 	else if (n == 0)
 	{
-		printf("%d is zero\n", n);
+		printf("%ld is zero\n", n);
 	}
 	else
 	{
-		printf("%d is negative\n", n);
+		printf("%ld is negative\n", n);
 	}
 	return (0);
 }
